website/requirements.txt
fastapi[all]
uvicorn
pymongo
motor
python-dotenv
mongoengine
bcrypt
psycopg2-binary
pyjwt